Bit checks

+++ assembler/src/ast.rs
@@ -53,7 +53,7 @@ pub enum Literal {
 
 #[derive(Debug, Clone)]
 pub enum Expression {
-    Number(i64),
+    Number(i32),
     Ident(String),
     Pos(Box<Spanned<Expression>>),
     Neg(Box<Spanned<Expression>>),
